Move page() helper out of mozorg and into base app
